requirements.txt
django==2.0.4
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-cors-headers==2.2.0
django-allauth==0.36.0
django-rest-auth ==0.9.3
django-friendship==1.5.0
drf-writable-nested==0.4.2
